finally fixed shader fucking errors.

+++ Engine/Included/2d_shared.glsl
@@ -8,7 +8,7 @@ uniform sampler2D DepthTex;
 uniform sampler2D DepthBlurTex;
 struct SpriteData {
     mat4 Transform;
-    uint Texture;
+    uint64_t Texture;
     float Depth;
     float Extra;
 };
